requirements.txt
keras==2.12.0
numpy==1.23.5
tensorflow-cpu==2.12.0
Flask==2.3.1
Flask-Cors==3.0.10
Pillow==9.5.0
streamlit==1.22.0